Define create_storage_key fixture in conftest.py
