+++ tests/test_fsm/storage/test_isolation.py
@@ -5,12 +5,6 @@ import pytest
 
 from aiogram.fsm.storage.base import BaseEventIsolation, StorageKey
 from aiogram.fsm.storage.redis import RedisEventIsolation, RedisStorage
-from tests.mocked_bot import MockedBot
-
-
-@pytest.fixture(name="storage_key")
-def create_storage_key(bot: MockedBot):
-    return StorageKey(chat_id=-42, user_id=42, bot_id=bot.id)
 
 
 @pytest.mark.parametrize(
