+++ tests/test_fsm/storage/test_mongo.py
@@ -2,16 +2,9 @@ import pytest
 
 from aiogram.fsm.state import State
 from aiogram.fsm.storage.mongo import MongoStorage, StorageKey
-from tests.mocked_bot import MockedBot
+from tests.conftest import CHAT_ID, USER_ID
 
 PREFIX = "fsm"
-CHAT_ID = -42
-USER_ID = 42
-
-
-@pytest.fixture(name="storage_key")
-def create_storage_key(bot: MockedBot):
-    return StorageKey(chat_id=CHAT_ID, user_id=USER_ID, bot_id=bot.id)
 
 
 async def test_update_not_existing_data_with_empty_dictionary(
